+++ src-docs/src/views/inline_edit/inline_edit_example.js
@@ -30,6 +30,17 @@ const inlineEditTitleSnippet = `<EuiInlineEditTitle
   size="m"
 />`;
 
+import InlineEditControlled from './inline_edit_controlled';
+const inlineEditControlledSource = require('!!raw-loader!./inline_edit_controlled');
+const inlineEditControlledSnippet = `<EuiInlineEditText
+  inputAriaLabel="Edit title inline"
+  value={inlineEditValue}
+  onChange={inlineEditOnChange}
+  onCancel={(previousValue) => {
+  setInlineEditValue(previousValue);
+  }}
+/>`;
+
 import InlineEditPlaceholder from './inline_edit_placeholder';
 const inlineEditPlaceholderSource = require('!!raw-loader!./inline_edit_placeholder');
 const inlineEditPlaceholderSnippet = `<EuiInlineEditText
@@ -95,7 +106,6 @@ const inlineEditModePropsSnippet = `<EuiInlineEditText
 
 export const InlineEditExample = {
   title: 'Inline edit',
-  isBeta: true,
   intro: (
     <>
       <EuiText>
@@ -151,6 +161,30 @@ export const InlineEditExample = {
       props: { EuiInlineEditTitle },
       snippet: inlineEditTitleSnippet,
       playground: inlineEditTitleConfig,
+    },
+    {
+      title: 'Controlled values',
+      text: (
+        <>
+          <p>
+            To control the text value for inline edit, use the{' '}
+            <EuiCode>value</EuiCode> prop to continually pass a text value. An
+            <EuiCode>onChange</EuiCode> callback is required to receive and
+            update <EuiCode>value</EuiCode> based on user input. An{' '}
+            <EuiCode>onCancel</EuiCode> callback is required to reset the value
+            to the previous read mode text value.
+          </p>
+        </>
+      ),
+      source: [
+        {
+          type: GuideSectionTypes.TSX,
+          code: inlineEditControlledSource,
+        },
+      ],
+      demo: <InlineEditControlled />,
+      props: { EuiInlineEditText },
+      snippet: inlineEditControlledSnippet,
     },
     {
       title: 'Saving edited text',