+++ src/components/accordion/accordion.test.tsx
@@ -7,7 +7,6 @@
  */
 
 import React from 'react';
-import { fireEvent } from '@testing-library/react';
 import { mount } from 'enzyme';
 import { shouldRenderCustomStyles } from '../../test/internal';
 import { requiredProps } from '../../test/required_props';
@@ -76,6 +75,47 @@ describe('EuiAccordion', () => {
 
         expect(container.firstChild).toMatchSnapshot();
       });
+
+      describe('paddingSize', () => {
+        (['s', 'm', 'l'] as const).forEach((paddingSize) => {
+          it(paddingSize, () => {
+            const { container, getByTestSubject } = render(
+              <EuiAccordion
+                id={getId()}
+                buttonProps={{ paddingSize, 'data-test-subj': 'button' }}
+              />
+            );
+            expect(container.firstChild).toMatchSnapshot();
+            expect(getByTestSubject('button').className).toContain(paddingSize);
+          });
+        });
+      });
+
+      describe('arrow padding affordance', () => {
+        it('removes the padding next to the side the arrow is on', () => {
+          const { getByTestSubject } = render(
+            <EuiAccordion
+              id={getId()}
+              buttonProps={{ paddingSize: 'm', 'data-test-subj': 'button' }}
+              arrowDisplay="right"
+            />
+          );
+          expect(getByTestSubject('button').className).toContain('arrowRight');
+        });
+
+        it('does not remove any padding if no arrow is displayed', () => {
+          const { getByTestSubject } = render(
+            <EuiAccordion
+              id={getId()}
+              buttonProps={{ paddingSize: 'm', 'data-test-subj': 'button' }}
+              arrowDisplay="none"
+            />
+          );
+          const buttonClass = getByTestSubject('button').className;
+          expect(buttonClass).not.toContain('arrowRight');
+          expect(buttonClass).not.toContain('arrowLeft');
+        });
+      });
     });
 
     describe('buttonElement', () => {
@@ -110,6 +150,7 @@ describe('EuiAccordion', () => {
         );
 
         expect(container.firstChild).toMatchSnapshot();
+        expect(container.firstChild).not.toHaveAttribute('inert');
       });
     });
 
@@ -281,40 +322,6 @@ describe('EuiAccordion', () => {
       component.find('button').at(0).simulate('click');
 
       expect(childWrapper).toBe(document.activeElement);
-    });
-
-    it('sets tabbable children to `tabIndex={-1}` when accordions are closed', () => {
-      const { container } = render(
-        <EuiAccordion id={getId()}>
-          <button>tabbable item one</button>
-          <input type="text" placeholder="tabbable item two" />
-          <a href="#">tabbable item three</a>
-          <div tabIndex={0}>tabbable item four</div>
-        </EuiAccordion>
-      );
-      const children = container.querySelector('.euiAccordion__children')!;
-
-      expect(children.querySelectorAll('[tabindex="-1"]')).toHaveLength(4);
-      expect(children).toMatchSnapshot();
-    });
-
-    it('restores tabbable children on accordion open', () => {
-      const { container, getByTestSubject } = render(
-        <EuiAccordion
-          id={getId()}
-          buttonProps={{ 'data-test-subj': 'trigger' }}
-        >
-          <button>tabbable item one</button>
-          <input type="text" placeholder="tabbable item two" />
-          <a href="#">tabbable item three</a>
-          <div tabIndex={0}>tabbable item four</div>
-        </EuiAccordion>
-      );
-      fireEvent.click(getByTestSubject('trigger'));
-      const children = container.querySelector('.euiAccordion__children')!;
-
-      expect(children.querySelectorAll('[tabindex="-1"]')).toHaveLength(0);
-      expect(children).toMatchSnapshot();
     });
   });
 });