+++ src/components/combo_box/combo_box_input/combo_box_input.tsx
@@ -193,8 +193,16 @@ export class EuiComboBoxInput<T> extends Component<
 
     const pills = selectedOptions
       ? selectedOptions.map((option) => {
-          const { key, label, color, onClick, append, prepend, ...rest } =
-            option;
+          const {
+            key,
+            label,
+            color,
+            onClick,
+            append,
+            prepend,
+            truncationProps,
+            ...rest
+          } = option;
           const pillOnClose =
             isDisabled || singleSelection || onClick
               ? undefined