+++ src/components/popover/popover.tsx
@@ -164,6 +164,18 @@ export interface EuiPopoverProps extends PropsWithChildren, CommonProps {
    */
   repositionOnScroll?: boolean;
   /**
+   * By default, popovers will attempt to position themselves along the initial
+   * axis specified. If there is not enough room either vertically or horizontally
+   * however, the popover will attempt to reposition itself along the secondary
+   * cross axis if there is room there instead.
+   *
+   * If you do not not want this repositioning to occur (and it is acceptable for
+   * the popover to appear offscreen), set this to false to disable this behavior.
+   *
+   * @default true
+   */
+  repositionToCrossAxis?: boolean;
+  /**
    * Must be set to true if using `EuiDragDropContext` within a popover,
    * otherwise your nested drag & drop will have incorrect positioning
    */
@@ -281,6 +293,7 @@ export class EuiPopover extends Component<Props, State> {
   static defaultProps: Partial<PropsWithDefaults> = {
     isOpen: false,
     ownFocus: true,
+    repositionToCrossAxis: true,
     anchorPosition: 'downCenter',
     panelPaddingSize: 'm',
     hasArrow: true,
@@ -482,7 +495,7 @@ export class EuiPopover extends Component<Props, State> {
   positionPopover = (allowEnforcePosition: boolean) => {
     if (this.button == null || this.panel == null) return;
 
-    const { anchorPosition } = this.props as PropsWithDefaults;
+    const { anchorPosition, offset = 0 } = this.props as PropsWithDefaults;
 
     let position = getPopoverPositionFromAnchorPosition(anchorPosition);
     let forcePosition = undefined;
@@ -507,15 +520,17 @@ export class EuiPopover extends Component<Props, State> {
       align: getPopoverAlignFromAnchorPosition(anchorPosition),
       anchor: this.button,
       popover: this.panel,
-      offset:
-        !this.props.attachToAnchor && this.props.hasArrow
-          ? 16 + (this.props.offset || 0)
-          : 8 + (this.props.offset || 0),
+      offset: this.props.attachToAnchor
+        ? offset
+        : this.props.hasArrow
+        ? 16 + offset
+        : 8 + offset,
       arrowConfig: {
         arrowWidth: 24,
         arrowBuffer: 10,
       },
       returnBoundingBox: this.props.attachToAnchor,
+      allowCrossAxis: this.props.repositionToCrossAxis,
       buffer: this.props.buffer,
     });
 
@@ -605,6 +620,7 @@ export class EuiPopover extends Component<Props, State> {
       hasArrow,
       arrowChildren,
       repositionOnScroll,
+      repositionToCrossAxis,
       hasDragDrop,
       zIndex,
       attachToAnchor,